refactor(composition): split test modules
