+++ crates/client/src/engine/composition/tests.rs
@@ -1,0 +1,60 @@
+use super::{
+    Candidates, ClauseActionBackend, ClauseActionStateMut, Composition, CompositionState,
+    TextServiceFactory,
+};
+use crate::engine::{
+    client_action::{ClientAction, SetTextType},
+    user_action::{Function, Navigation, UserAction},
+};
+use shared::{get_default_romaji_rows, AppConfig, RomajiRule, WidthMode};
+
+pub(super) fn row(input: &str, output: &str, next_input: &str) -> RomajiRule {
+    RomajiRule {
+        input: input.to_string(),
+        output: output.to_string(),
+        next_input: next_input.to_string(),
+    }
+}
+
+pub(super) fn candidates(
+    texts: &[&str],
+    sub_texts: &[&str],
+    hiragana: &str,
+    corresponding_count: &[i32],
+) -> Candidates {
+    Candidates {
+        texts: texts.iter().map(|value| (*value).to_string()).collect(),
+        sub_texts: sub_texts.iter().map(|value| (*value).to_string()).collect(),
+        hiragana: hiragana.to_string(),
+        corresponding_count: corresponding_count.to_vec(),
+    }
+}
+
+pub(super) fn actual_future_snapshot(
+    clause_preview: &str,
+    suffix: &str,
+    raw_input: &str,
+    raw_hiragana: &str,
+    corresponding_count: i32,
+) -> super::FutureClauseSnapshot {
+    TextServiceFactory::build_future_clause_snapshot(
+        clause_preview,
+        suffix,
+        raw_input,
+        raw_hiragana,
+        "",
+        corresponding_count,
+        0,
+        &candidates(
+            &[clause_preview],
+            &[suffix],
+            raw_hiragana,
+            &[corresponding_count],
+        ),
+    )
+}
+
+mod integration_patterns;
+mod snapshot_restore;
+pub(super) mod stateful_harness;
+mod symbol_and_width;
